requirements.txt
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pgvector==0.2.5
SQLAlchemy==2.0.36
pandas==2.2.3
numpy==1.26.4
sentence-transformers==3.1.1
langchain==0.3.6
langchain-community==0.3.4
groq==0.11.0
streamlit==1.38.0
tqdm==4.66.5